Onine Banking System
In"this venture we can keep record of every day banking exchanges. This undertaking is fit for holding any quantities of records. In first choice we can see record of a specific individual by giving essentially account number of that individual. In subsequent choice we can see the rundown all things considered. Through third alternative we can do banking exchanges (store/pull back) in fourth choice we can open new account. In the fifth choice we can alter or erase any record and last choice is stopped to leave the undertaking.”
The project maintains multiple accounts and customer details on-line. The project also maintains details for various account number and password assigned to them. A customer can login by entering account number and password to perform various types of transactions such as view the balance, deposit the amount and withdraw the amount and other activities.